Split remove WS app by global and context
