+++ backend/src/Equinor.ProjectExecutionPortal.WebApi/ViewModels/WorkSurfaceApp/ApiRemoveWorkSurfaceAppRequest.cs
@@ -1,12 +1,18 @@
-﻿using Equinor.ProjectExecutionPortal.Application.Commands.WorkSurfaces.RemoveWorkSurfaceApp;
+﻿using Equinor.ProjectExecutionPortal.Application.Commands.WorkSurfaces.RemoveContextWorkSurfaceApp;
+using Equinor.ProjectExecutionPortal.Application.Commands.WorkSurfaces.RemoveGlobalWorkSurfaceApp;
 
 namespace Equinor.ProjectExecutionPortal.WebApi.ViewModels.WorkSurfaceApp
 {
     public class ApiRemoveWorkSurfaceAppRequest
     {
-        public RemoveWorkSurfaceAppCommand ToCommand(Guid workSurfaceId, string? contextExternalId, string appKey)
+        public RemoveContextWorkSurfaceAppCommand ToCommand(Guid workSurfaceId, string contextExternalId, string appKey)
         {
-            return new RemoveWorkSurfaceAppCommand(workSurfaceId, contextExternalId, appKey);
+            return new RemoveContextWorkSurfaceAppCommand(workSurfaceId, contextExternalId, appKey);
+        }
+
+        public RemoveGlobalWorkSurfaceAppCommand ToCommand(Guid workSurfaceId, string appKey)
+        {
+            return new RemoveGlobalWorkSurfaceAppCommand(workSurfaceId, appKey);
         }
     }
 }
